refactor: move icon sections to an outside constant
