+++ src/components/IconsList.vue
@@ -1,9 +1,8 @@
 <script setup lang="ts">
-import { ref, computed } from 'vue';
+import { ref } from 'vue';
 import { VClosePopper as vClosePopper } from 'floating-vue';
 import CCIcon from './CCIcon.vue';
-import { EMOJIS } from '@/constants/emojis';
-import { ICONS } from '@/constants/hicons';
+import { SECTIONS } from '@/constants/sections'
 
 defineProps<{
   selectedIcon?: string
@@ -13,26 +12,7 @@ const emit = defineEmits<{
   (e: 'select', value: string): void
 }>()
 
-
-const emojsSections = computed(() => {
-  const hIconSection = {
-    name: 'Charts',
-    label: 'Charts',
-    icons: ICONS.map(icon => icon.icon)
-  }
-  const emojsSections = Object.keys(EMOJIS).map(key => {
-    const icons = Object.values(EMOJIS[key])
-    return {
-      name: key,
-      label: icons[0],
-      icons
-    }
-  })
-
-  return [hIconSection, ...emojsSections]
-})
-
-const activeSection = ref(emojsSections.value[0].name)
+const activeSection = ref(SECTIONS[0].name)
 
 function select(icon: string) {
   emit('select', icon)
@@ -44,7 +24,7 @@ function select(icon: string) {
     <!-- header -->
     <div class="tw-flex tw-justify-between tw-border-b tw-px-1">
       <div
-        v-for="section in emojsSections"
+        v-for="section in SECTIONS"
         :key="section.name"
         class="tw-cursor-pointer tw-p-1"
         :class="{ 'tw-border-blue-600 tw-border-b-2': activeSection === section.name }"
@@ -55,7 +35,7 @@ function select(icon: string) {
 
     <!-- body -->
     <div class="tw-max-h-80 tw-overflow-auto tw-p-2">
-      <div v-for="section in emojsSections" :key="section.name">
+      <div v-for="section in SECTIONS" :key="section.name">
         <div class="tw-font-semibold">
           {{ section.name }}
         </div>
